Move method to separated class
