+++ src/main/scala/nl/biopet/utils/biowdl/PipelineSuccess.scala
@@ -21,6 +21,27 @@
 
 package nl.biopet.utils.biowdl
 
+import org.testng.annotations.Test
+import play.api.libs.json._
+
 trait PipelineSuccess extends Pipeline {
   logMustHave("workflow finished with status 'Succeeded'".r)
+
+  protected def parseFinalOutputs: JsObject = {
+    val start = logLines.indexWhere(line =>
+      line.matches(".*Workflow .* complete. Final Outputs:$")) + 1
+    val end = logLines.indexWhere(_.startsWith("["), start)
+    val jsonText = logLines.slice(start, end).mkString("\n")
+    Json.parse(jsonText) match {
+      case o: JsObject => o
+      case _           => throw new IllegalArgumentException("Outputs should be a object")
+    }
+  }
+
+  def expectedOutput: Map[String, JsValue]
+
+  @Test(dependsOnGroups = Array("parseLog"))
+  def testExpectedOutput(): Unit = {
+    parseFinalOutputs.value shouldBe expectedOutput
+  }
 }
